Major upgrade in killaura & Remove redundant numeric casts

+++ 1.8.9-Forge/src/main/java/net/ccbluex/liquidbounce/injection/forge/mixins/entity/MixinAbstractClientPlayer.java
@@ -68,7 +68,7 @@ public abstract class MixinAbstractClientPlayer extends MixinEntityPlayer
 			}
 
 			final int i = getItemInUseDuration();
-			float f1 = (float) i / 20.0f;
+			float f1 = i / 20.0f;
 			f1 = f1 > 1.0f ? 1.0f : f1 * f1;
 			newFOV *= 1.0f - f1 * 0.15f;
 			callbackInfoReturnable.setReturnValue(newFOV);
